Added burgerking scraper to the django app

+++ accounts/mailing.py
@@ -13,19 +13,26 @@ def send_mail_task(name, email):
     print(to)
     email = EmailMessage(subject, body, 'shivamsri7011@gmail.com',['shivamsri701@gmail.com'])
     email.content_subtype = 'html'
+
+    stores = None
     if name == 'PizzaHut':
       stores = Store.objects.filter(name='Pizza Hut').values_list()
     elif name == 'Tacobell':
       stores = Store.objects.filter(name='Taco Bell').values_list()
     elif name == 'Starbucks':
       stores = Store.objects.filter(name='Starbucks').values_list()
+    elif name == 'Verizon':
+      stores = Store.objects.filter(name__icontains='Verizon').values_list()
+    elif name == 'Burgerking':
+      stores = Store.objects.filter(name='Burgerking').values_list()[:2000]
+    print('Stores mila', len(stores))
+    if stores is not None:
+      with open('accounts/{0}.csv'.format(name), 'w') as file:
+          writer = csv.writer(file)
+          writer.writerows(stores)
 
-    with open('accounts/{0}.csv'.format(name), 'w') as file:
-        writer = csv.writer(file)
-        writer.writerows(stores)
-
-    file = open('accounts/{0}.csv'.format(name), 'r')
-    email.attach('{0}.csv'.format(name), file.read(), 'text/plain')
-    email.send()
-    os.remove('accounts/{0}.csv'.format(name))
-    return None
+      file = open('accounts/{0}.csv'.format(name), 'r')
+      email.attach('{0}.csv'.format(name), file.read(), 'text/plain')
+      email.send()
+      os.remove('accounts/{0}.csv'.format(name))
+      return None
